debugging and removing unused variables

+++ dimensions.f90
@@ -2,6 +2,6 @@ module dimensions
       implicit none
       save
       integer, parameter:: nx = 202, ny = 3, nz = 400 
-      integer*4, parameter:: Ni_max =2000000
+      integer, parameter:: Ni_max =2000000
            
 end module dimensions
